Added logic, check functionality
