+++ frontend/src/App.js
@@ -1,109 +1,23 @@
-import React, { useState } from 'react';
+import React from "react";
+import { BrowserRouter as Router, Route, Routes, Link } from "react-router-dom";
+import Login from "./pages/Login";
+import Register from "./pages/Register";
+import Dashboard from "./pages/Dashboard";
 
 function App() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [message, setMessage] = useState('');
-  const [registerUsername, setRegisterUsername] = useState('');
-  const [registerPassword, setRegisterPassword] = useState('');
-  const [registerMessage, setRegisterMessage] = useState('');
-
-  const handleLogin = async (event) => {
-    event.preventDefault();
-    try {
-      const response = await fetch(`${process.env.REACT_APP_API_BASE_URL}/auth/login`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ username, password }),
-      });
-      const data = await response.json();
-      if (response.ok) {
-        setMessage(`Login successful! Token: ${data.token}`);
-      } else {
-        setMessage(data.message);
-      }
-    } catch (error) {
-      setMessage('Login failed');
-    }
-  };
-
-  const handleRegister = async (event) => {
-    event.preventDefault();
-    try {
-      const response = await fetch(`${process.env.REACT_APP_API_BASE_URL}/auth/register`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ username: registerUsername, password: registerPassword }),
-      });
-      const data = await response.json();
-      if (response.ok) {
-        setRegisterMessage('Registration successful!');
-      } else {
-        setRegisterMessage(data.message);
-      }
-    } catch (error) {
-      setRegisterMessage('Registration failed');
-    }
-  };
-
   return (
-    <div className="App">
-      <h1>Login</h1>
-      <form onSubmit={handleLogin}>
-        <div>
-          <label>
-            Username:
-            <input
-              type="text"
-              value={username}
-              onChange={(e) => setUsername(e.target.value)}
-            />
-          </label>
-        </div>
-        <div>
-          <label>
-            Password:
-            <input
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-          </label>
-        </div>
-        <button type="submit">Login</button>
-      </form>
-      {message && <h2>{message}</h2>}
+    <Router>
+      <h1>Welcome to Mindmates!</h1>
+      <nav>
+        <Link to="/">Login</Link> | <Link to="/register">Register</Link> | <Link to="/dashboard">Dashboard</Link>
+      </nav>
 
-      <h1>Register</h1>
-      <form onSubmit={handleRegister}>
-        <div>
-          <label>
-            Username:
-            <input
-              type="text"
-              value={registerUsername}
-              onChange={(e) => setRegisterUsername(e.target.value)}
-            />
-          </label>
-        </div>
-        <div>
-          <label>
-            Password:
-            <input
-              type="password"
-              value={registerPassword}
-              onChange={(e) => setRegisterPassword(e.target.value)}
-            />
-          </label>
-        </div>
-        <button type="submit">Register</button>
-      </form>
-      {registerMessage && <h2>{registerMessage}</h2>}
-    </div>
+      <Routes>
+        <Route path="/" element={<Login />} />
+        <Route path="/register" element={<Register />} />
+        <Route path="/dashboard" element={<Dashboard />} />
+      </Routes>
+    </Router>
   );
 }
