+++ frontend/src/pages/Dashboard.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Dashboard() {
+  return <h1>Welcome to the Dashboard!</h1>;
+}
+
+export default Dashboard;
